Update Main.java
Alright so I got the change description option done, and wrote a couple of methods for printing the lists and finding the movie by name in the list. With those I believe we can easily change any of the parameters of the Movie object, if we so want to. I also added a couple lines to the first and second option so it prompts the user to hit enter. That way it doesn't just print the menu right away.