Update Main.java
Okay so I put all of our collective work that I could find on github together. I did some error handling in the add a movie option and finished the change the release date option. It appears all we have left to do is the keeping the coming list ordered by release date, and #5 on the technical requirements page of the project ("start showing movies in the theater with a given release date").